fix: tests failed - break out the formatDamagePart from RagItem
